refactor(deduplikator_autorow): split views.py (1177L) into views/ package
Powtórzenie splitu z dev (commit 917909206) z uwzględnieniem nowych
funkcji wprowadzonych na feature branchu:
- helpers.py: + _read_param, _scientist_id_to_autor_id, _resolve_autor_id,
  get_running_scan, _get_pending_candidates_for_main_autor,
  _get_next_candidate_group
- duplicates.py: + lastname_suggestions
- ignore.py: split ignore_author → ignore_scientist + ignore_autor;
  reset_ignored_authors → reset_ignored_scientists + reset_ignored_autorzy;
  + _trigger_rescan_after_reset
- merge.py: scal_autorow_view + delete_author (bez zmian funkcjonalnych)
- scan.py: start_scan_view, cancel_scan_view, scan_status_view
- export.py: download_duplicates_xlsx
- __init__.py: re-export wszystkich symboli (backward compat z urls.py i
  testami importującymi z deduplikator_autorow.views)

views.py usunięty.
